requirements.txt
altair==5.4.1 ; python_version >= "3.13" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.13" and python_version < "4.0"
bcrypt==4.2.0 ; python_version >= "3.13" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.13" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.13" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.13" and python_version < "4.0"
captcha==0.6.0 ; python_version >= "3.13" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.13" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.13" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.13" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.13" and python_version < "4.0" and platform_system == "Windows"
contourpy==1.3.0 ; python_version >= "3.13" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.13" and python_version < "4.0"
entrypoints==0.4 ; python_version >= "3.13" and python_version < "4.0"
extra-streamlit-components==0.1.71 ; python_version >= "3.13" and python_version < "4.0"
faker==30.8.2 ; python_version >= "3.13" and python_version < "4.0"
favicon==0.7.0 ; python_version >= "3.13" and python_version < "4.0"
fonttools==4.54.1 ; python_version >= "3.13" and python_version < "4.0"
gitdb==4.0.11 ; python_version >= "3.13" and python_version < "4.0"
gitpython==3.1.43 ; python_version >= "3.13" and python_version < "4.0"
htbuilder==0.6.2 ; python_version >= "3.13" and python_version < "4.0"
idna==3.10 ; python_version >= "3.13" and python_version < "4.0"
jinja2==3.1.4 ; python_version >= "3.13" and python_version < "4.0"
jsonschema-specifications==2024.10.1 ; python_version >= "3.13" and python_version < "4.0"
jsonschema==4.23.0 ; python_version >= "3.13" and python_version < "4.0"
kiwisolver==1.4.7 ; python_version >= "3.13" and python_version < "4.0"
lxml==5.3.0 ; python_version >= "3.13" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.13" and python_version < "4.0"
markdown==3.7 ; python_version >= "3.13" and python_version < "4.0"
markdownlit==0.0.7 ; python_version >= "3.13" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.13" and python_version < "4.0"
matplotlib==3.9.2 ; python_version >= "3.13" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.13" and python_version < "4.0"
more-itertools==10.5.0 ; python_version >= "3.13" and python_version < "4.0"
narwhals==1.13.3 ; python_version >= "3.13" and python_version < "4.0"
numpy==2.1.3 ; python_version >= "3.13" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.13" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.13" and python_version < "4.0"
pillow==11.0.0 ; python_version >= "3.13" and python_version < "4.0"
plotly==5.24.1 ; python_version >= "3.13" and python_version < "4.0"
prometheus-client==0.21.0 ; python_version >= "3.13" and python_version < "4.0"
protobuf==5.28.3 ; python_version >= "3.13" and python_version < "4.0"
pyarrow==18.0.0 ; python_version >= "3.13" and python_version < "4.0"
pydeck==0.9.1 ; python_version >= "3.13" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.13" and python_version < "4.0"
pyjwt==2.9.0 ; python_version >= "3.13" and python_version < "4.0"
pymdown-extensions==10.12 ; python_version >= "3.13" and python_version < "4.0"
pyparsing==3.2.0 ; python_version >= "3.13" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.13" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.13" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.13" and python_version < "4.0"
referencing==0.35.1 ; python_version >= "3.13" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.13" and python_version < "4.0"
rich==13.9.4 ; python_version >= "3.13" and python_version < "4.0"
rpds-py==0.21.0 ; python_version >= "3.13" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.13" and python_version < "4.0"
smmap==5.0.1 ; python_version >= "3.13" and python_version < "4.0"
soupsieve==2.6 ; python_version >= "3.13" and python_version < "4.0"
st-annotated-text==4.0.1 ; python_version >= "3.13" and python_version < "4.0"
st-theme==1.2.3 ; python_version >= "3.13" and python_version < "4.0"
streamlit-authenticator==0.4.1 ; python_version >= "3.13" and python_version < "4.0"
streamlit-camera-input-live==0.2.0 ; python_version >= "3.13" and python_version < "4.0"
streamlit-card==1.0.2 ; python_version >= "3.13" and python_version < "4.0"
streamlit-embedcode==0.1.2 ; python_version >= "3.13" and python_version < "4.0"
streamlit-extras==0.5.0 ; python_version >= "3.13" and python_version < "4.0"
streamlit-faker==0.0.3 ; python_version >= "3.13" and python_version < "4.0"
streamlit-image-coordinates==0.1.9 ; python_version >= "3.13" and python_version < "4.0"
streamlit-keyup==0.2.4 ; python_version >= "3.13" and python_version < "4.0"
streamlit-toggle-switch==1.0.2 ; python_version >= "3.13" and python_version < "4.0"
streamlit-vertical-slider==2.5.5 ; python_version >= "3.13" and python_version < "4.0"
streamlit==1.40.0 ; python_version >= "3.13" and python_version < "4.0"
tenacity==9.0.0 ; python_version >= "3.13" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.13" and python_version < "4.0"
tornado==6.4.1 ; python_version >= "3.13" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.13" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.13" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.13" and python_version < "4.0"
validators==0.34.0 ; python_version >= "3.13" and python_version < "4.0"
watchdog==5.0.3 ; python_version >= "3.13" and python_version < "4.0" and platform_system != "Darwin"
